# packages/radar-contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InteractiveAgentTurnClaimRead": {
        "properties": {
          "claimedAt": {
            "format": "date-time",
            "title": "Claimedat",
            "type": "string"
          },
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "deviceId": {
            "format": "uuid",
            "title": "Deviceid",
            "type": "string"
          },
          "expiredAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiredat"
          },
          "failedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failedat"
          },
          "failureCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failurecode"
          },
          "heartbeatAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeatat"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "leaseExpiresAt": {
            "format": "date-time",
            "title": "Leaseexpiresat",
            "type": "string"
          },
          "localSessionId": {
            "format": "uuid",
            "title": "Localsessionid",
            "type": "string"
          },
          "lockVersion": {
            "minimum": 1.0,
            "title": "Lockversion",
            "type": "integer"
          },
          "modelAlias": {
            "title": "Modelalias",
            "type": "string"
          },
          "policyVersion": {
            "title": "Policyversion",
            "type": "string"
          },
          "requestCount": {
            "minimum": 0.0,
            "title": "Requestcount",
            "type": "integer"
          },
          "runtimeVersion": {
            "title": "Runtimeversion",
            "type": "string"
          },
          "schemaVersion": {
            "minimum": 1.0,
            "title": "Schemaversion",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/InteractiveAgentTurnStatus"
          },
          "turnToken": {
            "title": "Turntoken",
            "type": "string"
          }
        },
        "required": [
          "id",
          "localSessionId",
          "deviceId",
          "status",
          "runtimeVersion",
          "schemaVersion",
          "modelAlias",
          "policyVersion",
          "lockVersion",
          "requestCount",
          "leaseExpiresAt",
          "claimedAt",
          "turnToken"
        ],
        "title": "InteractiveAgentTurnClaimRead",
        "type": "object"
      },
      "InteractiveAgentTurnClaimRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotencyKey": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Idempotencykey",
            "type": "string"
          },
          "localSessionId": {
            "format": "uuid",
            "title": "Localsessionid",
            "type": "string"
          }
        },
        "required": [
          "localSessionId",
          "idempotencyKey"
        ],
        "title": "InteractiveAgentTurnClaimRequest",
        "type": "object"
      },
      "InteractiveAgentTurnCompleteRequest": {
        "additionalProperties": false,
        "properties": {
          "expectedLockVersion": {
            "minimum": 1.0,
            "title": "Expectedlockversion",
            "type": "integer"
          }
        },
        "required": [
          "expectedLockVersion"
        ],
        "title": "InteractiveAgentTurnCompleteRequest",
        "type": "object"
      },
      "InteractiveAgentTurnFailRequest": {
        "additionalProperties": false,
        "properties": {
          "expectedLockVersion": {
            "minimum": 1.0,
            "title": "Expectedlockversion",
            "type": "integer"
          },
          "failureCode": {
            "pattern": "^[a-z][a-z0-9_.-]{0,63}$",
            "title": "Failurecode",
            "type": "string"
          }
        },
        "required": [
          "expectedLockVersion",
          "failureCode"
        ],
        "title": "InteractiveAgentTurnFailRequest",
        "type": "object"
      },
      "InteractiveAgentTurnHeartbeatRequest": {
        "additionalProperties": false,
        "properties": {
          "expectedLockVersion": {
            "minimum": 1.0,
            "title": "Expectedlockversion",
            "type": "integer"
          }
        },
        "required": [
          "expectedLockVersion"
        ],
        "title": "InteractiveAgentTurnHeartbeatRequest",
        "type": "object"
      },
      "InteractiveAgentTurnRead": {
        "properties": {
          "claimedAt": {
            "format": "date-time",
            "title": "Claimedat",
            "type": "string"
          },
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "deviceId": {
            "format": "uuid",
            "title": "Deviceid",
            "type": "string"
          },
          "expiredAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiredat"
          },
          "failedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failedat"
          },
          "failureCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failurecode"
          },
          "heartbeatAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeatat"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "leaseExpiresAt": {
            "format": "date-time",
            "title": "Leaseexpiresat",
            "type": "string"
          },
          "localSessionId": {
            "format": "uuid",
            "title": "Localsessionid",
            "type": "string"
          },
          "lockVersion": {
            "minimum": 1.0,
            "title": "Lockversion",
            "type": "integer"
          },
          "modelAlias": {
            "title": "Modelalias",
            "type": "string"
          },
          "policyVersion": {
            "title": "Policyversion",
            "type": "string"
          },
          "requestCount": {
            "minimum": 0.0,
            "title": "Requestcount",
            "type": "integer"
          },
          "runtimeVersion": {
            "title": "Runtimeversion",
            "type": "string"
          },
          "schemaVersion": {
            "minimum": 1.0,
            "title": "Schemaversion",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/InteractiveAgentTurnStatus"
          }
        },
        "required": [
          "id",
          "localSessionId",
          "deviceId",
          "status",
          "runtimeVersion",
          "schemaVersion",
          "modelAlias",
          "policyVersion",
          "lockVersion",
          "requestCount",
          "leaseExpiresAt",
          "claimedAt"
        ],
        "title": "InteractiveAgentTurnRead",
        "type": "object"
      },
      "InteractiveAgentTurnStatus": {
        "enum": [
          "claimed",
          "running",
          "completed",
          "failed",
          "expired"
        ],
        "title": "InteractiveAgentTurnStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Opportunity IM Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/agent/interactive/turns": {
      "post": {
        "operationId": "claim_interactive_agent_turn_api_v1_agent_interactive_turns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractiveAgentTurnClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractiveAgentTurnClaimRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Claim Interactive Agent Turn",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/interactive/turns/{turn_id}/complete": {
      "post": {
        "operationId": "complete_interactive_agent_turn_api_v1_agent_interactive_turns__turn_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractiveAgentTurnCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractiveAgentTurnRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Interactive Agent Turn",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/interactive/turns/{turn_id}/expire": {
      "post": {
        "operationId": "expire_interactive_agent_turn_api_v1_agent_interactive_turns__turn_id__expire_post",
        "parameters": [
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractiveAgentTurnRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Expire Interactive Agent Turn",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/interactive/turns/{turn_id}/fail": {
      "post": {
        "operationId": "fail_interactive_agent_turn_api_v1_agent_interactive_turns__turn_id__fail_post",
        "parameters": [
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractiveAgentTurnFailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractiveAgentTurnRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fail Interactive Agent Turn",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/agent/interactive/turns/{turn_id}/heartbeat": {
      "post": {
        "operationId": "heartbeat_interactive_agent_turn_api_v1_agent_interactive_turns__turn_id__heartbeat_post",
        "parameters": [
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractiveAgentTurnHeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractiveAgentTurnRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Heartbeat Interactive Agent Turn",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}